% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evds_data2.R
\name{evds_data2}
\alias{evds_data2}
\title{Importing data from CBRT EVDS}
\usage{
evds_data2(anahtar, verisetleri, baslangic_tarihi, bitis_tarihi, frekans = "")
}
\arguments{
\item{anahtar}{Your API key from EVDS, sign up to EVDS for API key.}

\item{verisetleri}{Series you want to import.}

\item{baslangic_tarihi}{Start date you want}

\item{bitis_tarihi}{End date for the data you want}

\item{frekans}{Frequency for the data. Default is set automatically by EVDS.}
}
\description{
Importing data from CBRT EVDS
}
